Reentry into regionManager.activate handled.
If the activation of a part results in another set of parts being
activated (e.g. to navigate to an error page), then that error
page would not be shown as the initial activation would override
the just set parts. This rentry will now not perform final activation
of original parts now.