Properties now always have correct isValid state
Properties that have been made observable will now have their isValid
property represent the current state of the value, no waiting has to
occur
before seeing this value.

In addition the 'model' parameter to rules has been removed, required to
avoid confusion when validating individual properties that do not have
a need for a model. All validators can still be used as before with a
change
in usage.